tree-wide: Improve organisation of the tree
Primarily, organise structs and enum definitions such that:

- The library data model is defined in src/lib.rs

- Base specification details are contained in src/nvme.rs

- MI specification details are contained in src/nvme/mi.rs

- The device implementation src/nvme/mi/dev.rs

In the process, drop the use of num-traits in favour of the Discriminant
trait.
