+++ Cargo.toml
@@ -12,8 +12,6 @@ heapless = "0.8.0"
 hmac = { version = "0.12.1", default-features = false }
 log = "0.4.22"
 mctp = { version = "0.2.0", default-features = false }
-num-derive = { version = "0.4", default-features = false }
-num-traits = { version = "0.2", default-features = false }
 sha2 = { version = "0.10.9", default-features = false }
 uuid = { version = "1.17.0", default-features = false }
